refactor(server): extract control-plane wiring into pkg/server
Move the ~550-line wiring body and helpers out of cmd/vibed/main.go into a new
exported pkg/server package (Run + NewLogger + the http/deploy/audit helpers).
cmd/vibed/main.go shrinks to a small bootstrap, and a second binary can import
pkg/server and reuse the whole control-plane wiring instead of forking main().

Behavior-preserving: same init ordering, defers, os.Exit behavior, and signal
handling. The audit-handler test moves with auditHandler into pkg/server.
go build/vet/test green.
